refactor: modularize app modes
